cleanup next package
