+++ packages/next/src/actions/payments.ts
@@ -3,7 +3,7 @@
 import { revalidateTag } from 'next/cache';
 import type { Payment, CreatePaymentParams } from '@spree/sdk';
 import { getClient } from '../config';
-import { getCheckoutOptions, getCartId } from '../cookies';
+import { getCartOptions, requireCartId } from '../cookies';
 
 /**
  * Create a payment for a non-session payment method (e.g. Check, Cash on Delivery, Bank Transfer).
@@ -12,9 +12,8 @@ import { getCheckoutOptions, getCartId } from '../cookies';
 export async function createPayment(
   params: CreatePaymentParams
 ): Promise<Payment> {
-  const options = await getCheckoutOptions();
-  const cartId = await getCartId();
-  if (!cartId) throw new Error('No cart found');
+  const options = await getCartOptions();
+  const cartId = await requireCartId();
   const result = await getClient().carts.payments.create(cartId, params, options);
   revalidateTag('checkout');
   return result;
