+++ packages/next/tests/actions/cart.test.ts
@@ -8,11 +8,17 @@ const mockClient = {
     get: vi.fn(),
     list: vi.fn(),
     create: vi.fn(),
+    update: vi.fn(),
+    complete: vi.fn(),
     associate: vi.fn(),
     items: {
       create: vi.fn(),
       update: vi.fn(),
       delete: vi.fn(),
+    },
+    shipments: {
+      list: vi.fn(),
+      update: vi.fn(),
     },
     couponCodes: {
       apply: vi.fn(),
@@ -25,7 +31,10 @@ vi.mock('@spree/sdk', () => ({
   createClient: vi.fn(() => mockClient),
 }));
 
-import { getCart, getOrCreateCart, addItem, updateItem, removeItem, clearCart, associateCart } from '../../src/actions/cart';
+import {
+  getCart, getOrCreateCart, addItem, updateItem, removeItem, clearCart, associateCart,
+  updateCart, getShipments, selectShippingRate, applyCoupon, removeCoupon, complete,
+} from '../../src/actions/cart';
 import { revalidateTag } from 'next/cache';
 
 // Cookie names used by the actions:
@@ -219,6 +228,134 @@ describe('cart actions', () => {
         'cart_1',
         { spreeToken: 'cart_token', token: 'jwt_token' }
       );
+    });
+  });
+
+  describe('updateCart', () => {
+    it('updates cart and revalidates', async () => {
+      const updatedCart = { id: 'cart_1', number: 'R123' };
+      const params = { email: 'test@example.com' };
+      mockCookies({
+        '_spree_cart_token': 'order_token',
+        '_spree_cart_token_id': 'cart_1',
+        '_spree_jwt': 'jwt_token',
+      });
+      mockClient.carts.update.mockResolvedValue(updatedCart);
+
+      const result = await updateCart(params);
+      expect(result).toEqual(updatedCart);
+      expect(mockClient.carts.update).toHaveBeenCalledWith(
+        'cart_1',
+        params,
+        { spreeToken: 'order_token', token: 'jwt_token' }
+      );
+      expect(revalidateTag).toHaveBeenCalledWith('checkout');
+    });
+  });
+
+  describe('getShipments', () => {
+    it('returns shipments for the cart', async () => {
+      const mockShipments = { data: [{ id: 's1', shipping_rates: [] }] };
+      mockCookies({
+        '_spree_cart_token': 'order_token',
+        '_spree_cart_token_id': 'cart_1',
+        '_spree_jwt': 'jwt_token',
+      });
+      mockClient.carts.shipments.list.mockResolvedValue(mockShipments);
+
+      const result = await getShipments();
+      expect(result).toEqual(mockShipments);
+      expect(mockClient.carts.shipments.list).toHaveBeenCalledWith(
+        'cart_1',
+        { spreeToken: 'order_token', token: 'jwt_token' }
+      );
+    });
+  });
+
+  describe('selectShippingRate', () => {
+    it('selects shipping rate and revalidates', async () => {
+      const updatedCart = { id: 'cart_1', ship_total: '10.0' };
+      mockCookies({
+        '_spree_cart_token': 'order_token',
+        '_spree_cart_token_id': 'cart_1',
+        '_spree_jwt': 'jwt_token',
+      });
+      mockClient.carts.shipments.update.mockResolvedValue(updatedCart);
+
+      const result = await selectShippingRate('s1', 'sr1');
+      expect(result).toEqual(updatedCart);
+      expect(mockClient.carts.shipments.update).toHaveBeenCalledWith(
+        'cart_1',
+        's1',
+        { selected_shipping_rate_id: 'sr1' },
+        { spreeToken: 'order_token', token: 'jwt_token' }
+      );
+      expect(revalidateTag).toHaveBeenCalledWith('checkout');
+    });
+  });
+
+  describe('applyCoupon', () => {
+    it('applies coupon and revalidates checkout and cart', async () => {
+      const updatedCart = { id: 'cart_1', promo_total: -10 };
+      mockCookies({
+        '_spree_cart_token': 'order_token',
+        '_spree_cart_token_id': 'cart_1',
+        '_spree_jwt': 'jwt_token',
+      });
+      mockClient.carts.couponCodes.apply.mockResolvedValue(updatedCart);
+
+      const result = await applyCoupon('SAVE10');
+      expect(result).toEqual(updatedCart);
+      expect(mockClient.carts.couponCodes.apply).toHaveBeenCalledWith(
+        'cart_1',
+        'SAVE10',
+        { spreeToken: 'order_token', token: 'jwt_token' }
+      );
+      expect(revalidateTag).toHaveBeenCalledWith('checkout');
+      expect(revalidateTag).toHaveBeenCalledWith('cart');
+    });
+  });
+
+  describe('removeCoupon', () => {
+    it('removes coupon and revalidates checkout and cart', async () => {
+      const updatedCart = { id: 'cart_1', promo_total: 0 };
+      mockCookies({
+        '_spree_cart_token': 'order_token',
+        '_spree_cart_token_id': 'cart_1',
+        '_spree_jwt': 'jwt_token',
+      });
+      mockClient.carts.couponCodes.remove.mockResolvedValue(updatedCart);
+
+      const result = await removeCoupon('SAVE10');
+      expect(result).toEqual(updatedCart);
+      expect(mockClient.carts.couponCodes.remove).toHaveBeenCalledWith(
+        'cart_1',
+        'SAVE10',
+        { spreeToken: 'order_token', token: 'jwt_token' }
+      );
+      expect(revalidateTag).toHaveBeenCalledWith('checkout');
+      expect(revalidateTag).toHaveBeenCalledWith('cart');
+    });
+  });
+
+  describe('complete', () => {
+    it('completes checkout and revalidates', async () => {
+      const mockOrder = { id: 'cart_1', state: 'complete' };
+      mockCookies({
+        '_spree_cart_token': 'order_token',
+        '_spree_cart_token_id': 'cart_1',
+        '_spree_jwt': 'jwt_token',
+      });
+      mockClient.carts.complete.mockResolvedValue(mockOrder);
+
+      const result = await complete();
+      expect(result).toEqual(mockOrder);
+      expect(mockClient.carts.complete).toHaveBeenCalledWith(
+        'cart_1',
+        { spreeToken: 'order_token', token: 'jwt_token' }
+      );
+      expect(revalidateTag).toHaveBeenCalledWith('checkout');
+      expect(revalidateTag).toHaveBeenCalledWith('cart');
     });
   });
 });
